requirements.txt
pandas
numpy
openmeteo-requests
requests-cache
retry-requests
supabase
python-dotenv
scikit-learn
xgboost 
tensorflow 
joblib 
scikit-learn